Update index.md
Hi, thanks for making this site open to input. This site has been of great help to me, in trying to share with others in my business the kappa approach and the way that event sourcing works. Thank you :) :) 

There seems very few places that the principles are spelled out, and so after several years building event sourced systems, I spent the last two collating my learnings and also interviewed several other companies that had taken this approach. I put the results of my work into this talk, which I hope brings together the principles of each component, so that we get more common & shared understanding and can move forward faster. I hope you'll consider including it on this site, as I feel it'll be a help to anyone wanting to get a thorough overview or build their own system.

Once again, many thanks for compiling this information and making it available to everyone!